De-duplicate OMPException class
