+++ include/treelite/predictor.h
@@ -8,6 +8,7 @@
 #define TREELITE_PREDICTOR_H_
 
 #include <treelite/error.h>
+#include <treelite/omp_exception.h>
 #include <treelite/logging.h>
 #include <treelite/typeinfo.h>
 #include <treelite/c_api_runtime.h>
@@ -24,45 +25,6 @@
 
 namespace treelite {
 namespace predictor {
-
-/*!
- * \brief OMP Exception class catches, saves and rethrows exception from OMP blocks
- */
-class OMPException {
- private:
-  // exception_ptr member to store the exception
-  std::exception_ptr omp_exception_;
-  // mutex to be acquired during catch to set the exception_ptr
-  std::mutex mutex_;
-
- public:
-  /*!
-   * \brief Parallel OMP blocks should be placed within Run to save exception
-   */
-  template <typename Function, typename... Parameters>
-  void Run(Function f, Parameters... params) {
-    try {
-      f(params...);
-    } catch (treelite::Error &ex) {
-      std::lock_guard<std::mutex> lock(mutex_);
-      if (!omp_exception_) {
-        omp_exception_ = std::current_exception();
-      }
-    } catch (std::exception &ex) {
-      std::lock_guard<std::mutex> lock(mutex_);
-      if (!omp_exception_) {
-        omp_exception_ = std::current_exception();
-      }
-    }
-  }
-
-  /*!
-   * \brief should be called from the main thread to rethrow the exception
-   */
-  void Rethrow() {
-    if (this->omp_exception_) std::rethrow_exception(this->omp_exception_);
-  }
-};
 
 /*! \brief data layout. The value -1 signifies the missing value.
     When the "missing" field is set to -1, the "fvalue" field is set to
