+++ src/CMakeLists.txt
@@ -142,6 +142,7 @@ target_sources(objtreelite_common
     ${PROJECT_SOURCE_DIR}/include/treelite/error.h
     ${PROJECT_SOURCE_DIR}/include/treelite/logging.h
     ${PROJECT_SOURCE_DIR}/include/treelite/math.h
+    ${PROJECT_SOURCE_DIR}/include/treelite/omp_exception.h
     ${PROJECT_SOURCE_DIR}/include/treelite/typeinfo.h
     ${PROJECT_SOURCE_DIR}/include/treelite/data.h
 )
